+++ tests/frontends/mpd/playback_test.py
@@ -325,6 +325,13 @@ class PlaybackControlHandlerTest(unittest.TestCase):
         result = self.h.handle_request(u'seek "1" "30"')
         self.assertEqual(self.b.playback.current_track, seek_track)
 
+    def test_seek_without_quotes(self):
+        self.b.current_playlist.append([Track(length=40000)])
+        self.h.handle_request(u'seek 0')
+        result = self.h.handle_request(u'seek 0 30')
+        self.assert_(u'OK' in result)
+        self.assert_(self.b.playback.time_position >= 30000)
+
     def test_seekid(self):
         self.b.current_playlist.append([Track(length=40000)])
         result = self.h.handle_request(u'seekid "0" "30"')